# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/restaurants": {
      "get": {
        "description": "Возвращает список ресторанов с поддержкой фильтрации:\n    - Поиск по названию\n    - Фильтр по городу и стране\n    - Фильтр по активности (по умолчанию только активные)\n    - Фильтр по минимальному рейтингу\n    Результаты отсортированы по рейтингу (лучшие сначала)",
        "operationId": "RestaurantsController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Поиск по названию ресторана",
            "schema": {
              "example": "тбилиси",
              "type": "string"
            }
          },
          {
            "name": "city",
            "required": false,
            "in": "query",
            "description": "Фильтр по городу",
            "schema": {
              "example": "Тбилиси",
              "type": "string"
            }
          },
          {
            "name": "country",
            "required": false,
            "in": "query",
            "description": "Фильтр по стране",
            "schema": {
              "example": "Грузия",
              "type": "string"
            }
          },
          {
            "name": "is_active",
            "required": false,
            "in": "query",
            "description": "Показывать только активные рестораны",
            "schema": {
              "default": true,
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "min_rating",
            "required": false,
            "in": "query",
            "description": "Минимальный рейтинг",
            "schema": {
              "minimum": 0,
              "maximum": 5,
              "example": 4,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список ресторанов успешно получен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RestaurantResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Получить список ресторанов",
        "tags": [
          "Restaurants"
        ]
      },
      "post": {
        "description": "Создает новый ресторан в системе",
        "operationId": "RestaurantsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRestaurantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ресторан успешно создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestaurantResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные данные для создания ресторана"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Создать новый ресторан",
        "tags": [
          "Restaurants"
        ]
      }
    },
    "/restaurants/{id}": {
      "get": {
        "description": "Возвращает детальную информацию о ресторане включая:\n    - Основную информацию\n    - Статистику (количество столиков, вместимость, отзывы, средний рейтинг)",
        "operationId": "RestaurantsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Уникальный идентификатор ресторана",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ресторан найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestaurantDetailResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Ресторан с указанным ID не найден"
          }
        },
        "summary": "Получить ресторан по ID",
        "tags": [
          "Restaurants"
        ]
      },
      "patch": {
        "description": "Обновляет информацию о ресторане",
        "operationId": "RestaurantsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Уникальный идентификатор ресторана",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRestaurantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ресторан успешно обновлен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestaurantResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные данные для обновления ресторана"
          },
          "404": {
            "description": "Ресторан с указанным ID не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Обновить ресторан",
        "tags": [
          "Restaurants"
        ]
      },
      "delete": {
        "description": "Полностью удаляет ресторан из системы",
        "operationId": "RestaurantsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Уникальный идентификатор ресторана",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ресторан успешно удален"
          },
          "404": {
            "description": "Ресторан с указанным ID не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Удалить ресторан",
        "tags": [
          "Restaurants"
        ]
      }
    },
    "/restaurants/{id}/deactivate": {
      "patch": {
        "description": "Помечает ресторан как неактивный (is_active = false)",
        "operationId": "RestaurantsController_deactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Уникальный идентификатор ресторана",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ресторан успешно деактивирован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestaurantResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Ресторан с указанным ID не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Деактивировать ресторан",
        "tags": [
          "Restaurants"
        ]
      }
    },
    "/restaurants/{id}/activate": {
      "patch": {
        "description": "Помечает ресторан как активный (is_active = true)",
        "operationId": "RestaurantsController_activate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Уникальный идентификатор ресторана",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ресторан успешно активирован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestaurantResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Ресторан с указанным ID не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Активировать ресторан",
        "tags": [
          "Restaurants"
        ]
      }
    }
  },
  "info": {
    "title": "Georgian Restaurant API",
    "description": "API для управления грузинскими ресторанами с системой бронирования, заказов и аудита",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "RestaurantResponseDto": {
        "type": "object",
        "properties": {
          "restaurant_id": {
            "type": "number",
            "description": "Уникальный идентификатор ресторана",
            "example": 1
          },
          "restaurant_name": {
            "type": "string",
            "description": "Название ресторана",
            "example": "Ресторан \"Тбилиси\"",
            "minLength": 3,
            "maxLength": 100
          },
          "restaurant_description": {
            "type": "string",
            "description": "Описание ресторана",
            "example": "Аутентичная грузинская кухня в самом сердце города"
          },
          "country": {
            "type": "string",
            "description": "Страна",
            "example": "Грузия",
            "default": "Грузия"
          },
          "city": {
            "type": "string",
            "description": "Город",
            "example": "Тбилиси",
            "default": "Тбилиси"
          },
          "street_address": {
            "type": "string",
            "description": "Адрес ресторана",
            "example": "ул. Руставели, 15"
          },
          "working_hours": {
            "type": "object",
            "description": "Часы работы по дням недели",
            "example": {
              "monday": "10:00-22:00",
              "tuesday": "10:00-22:00",
              "wednesday": "10:00-22:00",
              "thursday": "10:00-22:00",
              "friday": "10:00-23:00",
              "saturday": "10:00-23:00",
              "sunday": "12:00-21:00"
            }
          },
          "is_active": {
            "type": "boolean",
            "description": "Активен ли ресторан",
            "example": true,
            "default": true
          },
          "rating": {
            "type": "number",
            "description": "Рейтинг ресторана",
            "example": 4.5,
            "minimum": 0,
            "maximum": 5
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "Дата создания",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "description": "Дата последнего обновления",
            "example": "2024-01-15T10:30:00.000Z"
          }
        },
        "required": [
          "restaurant_id",
          "restaurant_name",
          "country",
          "city",
          "street_address",
          "is_active",
          "rating",
          "created_at",
          "updated_at"
        ]
      },
      "RestaurantDetailResponseDto": {
        "type": "object",
        "properties": {
          "restaurant_id": {
            "type": "number",
            "description": "Уникальный идентификатор ресторана",
            "example": 1
          },
          "restaurant_name": {
            "type": "string",
            "description": "Название ресторана",
            "example": "Ресторан \"Тбилиси\"",
            "minLength": 3,
            "maxLength": 100
          },
          "restaurant_description": {
            "type": "string",
            "description": "Описание ресторана",
            "example": "Аутентичная грузинская кухня в самом сердце города"
          },
          "country": {
            "type": "string",
            "description": "Страна",
            "example": "Грузия",
            "default": "Грузия"
          },
          "city": {
            "type": "string",
            "description": "Город",
            "example": "Тбилиси",
            "default": "Тбилиси"
          },
          "street_address": {
            "type": "string",
            "description": "Адрес ресторана",
            "example": "ул. Руставели, 15"
          },
          "working_hours": {
            "type": "object",
            "description": "Часы работы по дням недели",
            "example": {
              "monday": "10:00-22:00",
              "tuesday": "10:00-22:00",
              "wednesday": "10:00-22:00",
              "thursday": "10:00-22:00",
              "friday": "10:00-23:00",
              "saturday": "10:00-23:00",
              "sunday": "12:00-21:00"
            }
          },
          "is_active": {
            "type": "boolean",
            "description": "Активен ли ресторан",
            "example": true,
            "default": true
          },
          "rating": {
            "type": "number",
            "description": "Рейтинг ресторана",
            "example": 4.5,
            "minimum": 0,
            "maximum": 5
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "Дата создания",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "description": "Дата последнего обновления",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "stats": {
            "type": "object",
            "description": "Статистика ресторана"
          }
        },
        "required": [
          "restaurant_id",
          "restaurant_name",
          "country",
          "city",
          "street_address",
          "is_active",
          "rating",
          "created_at",
          "updated_at"
        ]
      },
      "CreateRestaurantDto": {
        "type": "object",
        "properties": {
          "restaurant_name": {
            "type": "string",
            "description": "Название ресторана",
            "example": "Ресторан \"Тбилиси\"",
            "minLength": 3,
            "maxLength": 100
          },
          "restaurant_description": {
            "type": "string",
            "description": "Описание ресторана",
            "example": "Аутентичная грузинская кухня в самом сердце города"
          },
          "country": {
            "type": "string",
            "description": "Страна",
            "example": "Грузия",
            "default": "Грузия"
          },
          "city": {
            "type": "string",
            "description": "Город",
            "example": "Тбилиси",
            "default": "Тбилиси"
          },
          "street_address": {
            "type": "string",
            "description": "Адрес ресторана",
            "example": "ул. Руставели, 15",
            "minLength": 5
          },
          "working_hours": {
            "type": "object",
            "description": "Часы работы по дням недели",
            "example": {
              "monday": "10:00-22:00",
              "tuesday": "10:00-22:00",
              "wednesday": "10:00-22:00",
              "thursday": "10:00-22:00",
              "friday": "10:00-23:00",
              "saturday": "10:00-23:00",
              "sunday": "12:00-21:00"
            }
          },
          "is_active": {
            "type": "boolean",
            "description": "Активен ли ресторан",
            "example": true,
            "default": true
          }
        },
        "required": [
          "restaurant_name",
          "street_address"
        ]
      },
      "UpdateRestaurantDto": {
        "type": "object",
        "properties": {
          "restaurant_name": {
            "type": "string",
            "description": "Название ресторана",
            "example": "Ресторан \"Тбилиси\"",
            "minLength": 3,
            "maxLength": 100
          },
          "restaurant_description": {
            "type": "string",
            "description": "Описание ресторана",
            "example": "Аутентичная грузинская кухня в самом сердце города"
          },
          "country": {
            "type": "string",
            "description": "Страна",
            "example": "Грузия"
          },
          "city": {
            "type": "string",
            "description": "Город",
            "example": "Тбилиси"
          },
          "street_address": {
            "type": "string",
            "description": "Адрес ресторана",
            "example": "ул. Руставели, 15",
            "minLength": 5
          },
          "working_hours": {
            "type": "object",
            "description": "Часы работы по дням недели",
            "example": {
              "monday": "10:00-22:00",
              "tuesday": "10:00-22:00",
              "wednesday": "10:00-22:00",
              "thursday": "10:00-22:00",
              "friday": "10:00-23:00",
              "saturday": "10:00-23:00",
              "sunday": "12:00-21:00"
            }
          },
          "is_active": {
            "type": "boolean",
            "description": "Активен ли ресторан",
            "example": true
          },
          "rating": {
            "type": "number",
            "description": "Рейтинг ресторана (только для админов)",
            "example": 4.5,
            "minimum": 0,
            "maximum": 5
          }
        }
      }
    }
  }
}